+++ R/package.R
@@ -1,13 +1,4 @@
 
-#' Call R from R
-#'
-#' It is sometimes useful to perform a computation in a separate R
-#' process, without affecting the current R process at all. This packages
-#' does exactly that.
-#'
-#' @name callr
-"_PACKAGE"
-
 ## R CMD check workaround
 dummy_r6 <- function() R6::R6Class