+++ tests/testthat.R
@@ -1,4 +1,4 @@
 library(testthat)
 library(callr)
 
-if (Sys.getenv("NOT_CRAN", "") == "true") test_check("callr")
+test_check("callr")