+++ tests/testthat/test-eval.R
@@ -1,6 +1,4 @@
 
-context("r")
-
 test_that("basic r", {
   expect_equal(r(function() 1 + 1), 2)
   expect_equal(r(function(x) 1 + x, list(5)), 6)
@@ -247,6 +245,25 @@ test_that("local .Rprofile is not loaded from actual wd", {
   cat("aa <- 123\n", file = ".Rprofile")
   out <- callr::r(function() ls(.GlobalEnv), wd = wd2)
   expect_equal(out, character())
+})
+
+test_that("local .Rprofile is not loaded recursively", {
+  tmp <- tempfile()
+  on.exit(unlink(tmp, recursive = TRUE), add = TRUE)
+  dir.create(tmp)
+  wd <- getwd()
+  on.exit(setwd(wd), add = TRUE)
+  setwd(tmp)
+
+  expr <- quote({
+    rprofile <- Sys.getenv("R_PROFILE_USER", "~/.Rprofile")
+    if (file.exists(rprofile)) source(rprofile)
+    rm(rprofile)
+    aa <- 123
+  })
+  cat(deparse(expr), file = ".Rprofile", sep = "\n")
+  out <- callr::r(function() aa)
+  expect_equal(out, 123)
 })
 
 test_that("symbolic arguments are protected", {