+++ tests/testthat/test-r-process.R
@@ -1,6 +1,4 @@
 
-context("r_process")
-
 test_that("create and run r_process", {
   options <- r_process_options(func = function() 1 + 1)
   x <- r_process$new(options)